+++ pytest.ini
@@ -1,0 +1,5 @@
+[pytest]
+testpaths = tests
+filterwarnings =
+    ignore::DeprecationWarning
+
